+++ CHANGELOG.md
@@ -22,6 +22,8 @@ All notable changes to this project will be documented in this file.
     - Fixed #567: heatmap error (race condition) in PrinterClean for any newly created database
     - Fixed #576, #577: correct some function calls in PrinterClean
     - Fixed #542, #381: ensureIndex mongoose warning and circular Runner import resolved 
+    - Fixed #598: printer settingsAppearance missing will not cause failure anymore
+    - Fixed #596: changed OctoPrint plugin manager repository to new route with backwards compatibility version check 
 
 
 ## [v1.1.13-hotfix]